requirements.txt
python-dotenv
requests
google-genai
groq
pytest